test: add more cosign tests
