+++ tox.ini
@@ -1,6 +1,7 @@
 [vars]
 PACKAGE_MODULE = src/mobster
 PACKAGE_TESTS = tests
+PACKAGE_INTEGRATION_TESTS = tests/integration
 
 [tox]
 envlist =
@@ -25,13 +26,18 @@ setenv =
 
 [testenv:test]
 commands = pytest -v \
-       --ignore=tests/integration/ \
+       --ignore={[vars]PACKAGE_INTEGRATION_TESTS} \
        --cov {[vars]PACKAGE_MODULE} \
        --cov-report term-missing \
        --cov-fail-under 95 \
        --cov-report xml \
        --cov-report html \
        {[vars]PACKAGE_TESTS} \
+       {posargs}
+
+[testenv:test-integration]
+commands = pytest -v \
+       {[vars]PACKAGE_INTEGRATION_TESTS} \
        {posargs}
 
 [testenv:ruff]
